Feat: Adds handlers for updating the users password
Have added a new state to set the supabase event, such as PASSWORD_RECOVERY. Passing this through our context will let us handle the state throughout the different pages.

The sendResetPasswordLink will send a link to the user which directs them back to the resetPassword page. This page is loaded with the PASSWORD_RECOVERY event, allowing use to use that to toggle the form fields available.

updateUserPassword then updates the password and redirects the user back to the profile page along with the 'USER_UPDATED' event